Build a js bundle exposing diagram rendering as a library
The build of the composer web module now creates another bundle exposing the diagram rendering logic as a library and the vscode plugin renderer uses this.
