+++ tool-plugins/vscode/plugin/renderer/renderer.js
@@ -35,9 +35,9 @@ function renderDiagram(modelJson) {
         <meta charset="utf-8">
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="file://${__dirname}/resources/css/bundle.css" />
-        <link rel="stylesheet" href="file://${__dirname}/resources/css/default.css" />
-        <link rel="stylesheet" href="file://${__dirname}/resources/css/less.css" />
+        <link rel="stylesheet" href="file://${__dirname}/resources/bundle.css" />
+        <link rel="stylesheet" href="file://${__dirname}/resources/theme.css" />
+        <link rel="stylesheet" href="file://${__dirname}/resources/less.css" />
         <style>
             .overlay {
                 display: none;
@@ -57,7 +57,7 @@ function renderDiagram(modelJson) {
     <div class="ballerina-editor design-view-container" id="diagram">
     </div>
     </body>
-    <script charset="UTF-8" src="file://${__dirname}/resources/js/diagram.js"></script>
+    <script charset="UTF-8" src="file://${__dirname}/resources/ballerina-diagram-library.js"></script>
     <script>
             const json = window.modeljsonbal = ${jsonModel};
             window.onresize = () => {
